API: Update lib.polynomial and lib.npyio namespaces
