+++ numpy/core/tests/test_deprecations.py
@@ -750,7 +750,7 @@ class TestLibImports(_DeprecationTestCase):
     # Deprecated in Numpy 1.26.0, 2023-09
     def test_lib_functions_deprecation_call(self):
         from numpy.lib._utils_impl import safe_eval
-        from numpy.lib.npyio import recfromcsv, recfromtxt
+        from numpy.lib._npyio_impl import recfromcsv, recfromtxt
         from numpy.lib._function_base_impl import disp
         from numpy.lib._shape_base_impl import get_array_wrap
         from numpy.core.numerictypes import maximum_sctype
