+++ numpy/lib/_polynomial_impl.py
