+++ numpy/lib/_polynomial_impl.pyi
